trying to just use foreground in ios level and leave fg/bg compositing to uxn/devices/screen.c

+++ app/Emulator/Platform/emu_platform.h
@@ -18,7 +18,7 @@ typedef float r32;
 typedef struct {
     u16 width;
     u16 height;
-    u8* pixels;
+    u32* pixels;
 } PlatformBitmap;
